requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio
torchsummary
tqdm~=4.66.1
pytube~=15.0.0
openai-whisper